Add files via upload
Routing Information Protocol (RIP) is one of the oldest distance-vector routing protocols used to determine the best route for data packets within a network based on the distance to the destination. It uses hop count as its primary metric for path selection, where each hop represents a node or router the data must pass through to reach its destination.